refactor(runtime/llm): extract execution-envelope helpers to own module
Moves the ~120 LOC envelope enforcement + path-canonicalization
helpers out of chat-executor-tool-loop.ts into a new
chat-executor-envelope-helpers.ts. The main loop file drops from
2,342 -> 2,219 LOC.

Moved exports:
- getExecutionEnvelopeFilesystemAccessMode
- canonicalizeExplicitArtifactReferenceArgs
- enforceTopLevelExecutionEnvelope

Moved constants (all file-private in the new module):
- READ_ONLY_ENVELOPE_TOOL_NAMES
- WRITE_ENVELOPE_TOOL_MODES
- ENVELOPE_TOOL_PATH_ARG_KEYS

Removed now-unused imports from the main loop:
- ArtifactAccessMode (artifact-contract.js)
- ExecutionEnvelope (execution-envelope.js)
- isPathWithinAnyRoot / normalizeEnvelopePath / normalizeEnvelopeRoots /
  normalizeWorkspaceRoot / resolveExplicitArtifactReferencePath
  (path-normalization.js)
- getExecutionEnvelopeFilesystemAccessMode (now in the new module)

Purely file-organization — zero behavior delta. Typecheck green.
llm/ test suite (720 tests, 1 pre-existing token-budget compaction
failure unchanged from main) matches baseline.
